% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{b_spline}
\alias{b_spline}
\title{B spline basis with option of centering and differencing coefficients}
\usage{
b_spline(x, df = NULL, knots = NULL, degree = 3, intercept = FALSE,
  Boundary.knots = range(x), differences = TRUE, center = FALSE)
}
\arguments{
\item{center}{}
}
\description{
parameters follow \link{bs}
}
\details{
uses code from \link{bs}, but optionally transforms the basis into a difference
between spline parameters.

Also, optionally centers the spline so that the spline average is zero.
}
\references{
following code from https://github.com/SurajGupta/r-source/blob/master/src/library/splines/R/splines.R
}
