% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{bs_binomial_model}
\alias{bs_binomial_model}
\title{Sample Eta given B-splines individual-level model with binomial observations}
\usage{
bs_binomial_model(Y, data_model_parameters, BSFG_state = list())
}
\arguments{
\item{Y}{a vector of observation. Pre-scaled and centered if desired.}

\item{data_model_parameters}{a list including:
\code{observations}, a data.frame with columns: \code{ID}, \code{N} and \code{covariate}, ordered by ID.
and the variables df, knots, degree and intercept.
Empty values will use the defaults for \code{bs()}.}
}
\description{
Eta is a matrix of individual-level parmaters for a B-spline with
   equivalent knots over all individuals
}
\details{
This function should pre-calculate design matrices for each individual (assuming they are all
    unique)
}
