% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{bs_model}
\alias{bs_model}
\title{Sample Eta given B-splines individual-level model}
\usage{
bs_model(Y, data_model_parameters, BSFG_state = list())
}
\arguments{
\item{Y}{a vector of observation. Pre-scaled and centered if desired.}

\item{data_model_parameters}{a list including:
1) \code{observations}, a data.frame including columns: \code{ID}, and \code{covariate}, ordered by ID.
2) the variables df, knots, degree and intercept. Empty values will use the defaults for \code{bs()}
3) \code{individual_model} formula for each individual of form \code{~SPLINE:factor} using keyword
    \code{SPLINE} to denote b-spline factors. If not included, will use model \code{~SPLINE}}
}
\description{
Eta is a matrix of individual-level parmaters for a B-spline with
   equivalent knots over all individuals
}
\details{
This function should pre-calculate design matrices for each individual (assuming they are all
    unique), as well as SVDs for efficient repeated solvings of the posterior distributions
    as resid_Y_prec, Eta_mean and resid_Eta_prec are updated each iteraction.
}
