% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_control}
\alias{BSFG_control}
\title{Set BSFG run parameters}
\usage{
BSFG_control(sampler = c("fast_BSFG", "general_BSFG"), fixed_factors = c(T,
  F), simulation = c(F, T), scale_Y = c(T, F), b0 = 1, b1 = 5e-04,
  epsilon = 0.1, prop = 1, k_init = 20, h2_divisions = 100,
  burn = 100, thin = 2)
}
\arguments{
\item{sampler}{specify the sampler to use. fast_BSFG is much faster, but only allows one random
effect. If more are specified in \code{BSFG_init}, this is switched to general_BSFG.}

\item{fixed_factors}{should the fixed effect model be applied to factors as well as the factor
residuals? If so, the first column of the design matrix is dropped.}

\item{scale_Y}{Should the Y values be centered and scaled? Recommend, except for simulated data.}

\item{b0}{parameter of the \code{update_k} function. See Bhattacharya and Dunson 2011}

\item{b1}{parameter of the \code{update_k} function. See Bhattacharya and Dunson 2011}

\item{epsilon}{parameter of the \code{update_k} function. Smallest \lambda_{ij} that is
considered "large", signifying a factor should be kept}

\item{prop}{proportion of \lambda{ij} elements in a column of \Lambda that must be smaller than
\code{epsilon} before factor is dropped}

\item{h2_divisions}{A scalar or vector of length equal to number of random effects, specifying
the number of equally spaced divisions to devide each random effect variance component. All
variance componets are scaled as percent_of_total. This number of divisions are created for
each component, and then combined with \code{expand.grid()}. Then combinations of variances that sum to less than 1 are selected as valid.}

\item{simulaiton}{Is this a fit to simulated data? If so, a setup list will be expected providing
the true values}

\item{kinit}{initial number of factors}
}
\description{
Function to create run_parameters list for initializing BSFG model
}
\seealso{
\code{\link{BSFG_init}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}},
}
