% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_init}
\alias{BSFG_init}
\title{Initialize a BSFG model}
\usage{
BSFG_init(Y, fixed, random, data, priors, run_parameters, A_mats = NULL,
  A_inv_mats = NULL, fixed_Factors = NULL, scaleY = TRUE,
  sampler = "fast_BSFG", ncores = 1, simulation = F, setup = NULL,
  verbose = T)
}
\arguments{
\item{Y}{a n x p matrix of data (n individuals x p traits)}

\item{fixed}{formula for the fixed effects. \strong Note: currently only applies to residuals
(not factors)}

\item{random}{formula for the random effects. \strong Note: currently should be of the form \code{~ x1
+ x2} (no interactions)}

\item{data}{data.frame with n rows containing columns corresponding to the fixed and random
effects}

\item{priors}{list providing hyperparameters for the model priors. See details}

\item{run_parameters}{list providing various parameters for the model run}

\item{A_mats}{list of covariance matrices for random effects. If none provided (and none provided
for A_inv_mats), assumed to be the identity.}

\item{A_inv_mats}{list of covariance matrices for random effects. If none provided (and none
provided for A_mats), assumed to be the identity.}

\item{fixed_Factors}{currently no effect. To be used to specify fixed effect model for factors}

\item{scaleY}{should columns of Y be re-scaled to mean = 0, var = 1?}

\item{sampler}{('fast_BSFG','general_BSFG'). Specifies which Gibbs sampler to use. If more than
one random effect, defaults to general_BSFG, regardless of this parameter}

\item{ncores}{for \code{general_BSFG}, number of cores to use. \code{fast_BSFG} is parallelize by RcppParallel}

\item{simulation}{is this a simulation (and are known values included in setup?)?}

\item{setup}{a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}

\item{verbose}{(T,F) should progress in initialization be reported?}
}
\value{
An object of class BSFG_state with components:

current_state: a list of parameters in the current iteration of the sampler

Posterior: a list of arrays of posterior samples

RNG current state of R's Random number generator (for re-starting chaings)

traitnames: vector of trait names (from colnames of Y)

run_parameters, run_variables, data_matrices, priors, simulation
}
\description{
Sets up the BSFG model, selects starting values, and pre-calculates matrices for the GIBBS
sampler.
}
