% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{combined_model}
\alias{combined_model}
\title{Sample Eta given a list of different observation_models}
\usage{
combined_model(observation_model_parameters, BSFG_state = list())
}
\arguments{
\item{observation_model_parameters}{a list of observation_model lists}

\item{BSFG_state}{the current BSFG_state object. Can be just a list with \code{data}}
}
\description{
Sample Eta given a list of different observation_models
}
