% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{get_posterior_mean}
\alias{get_posterior_mean}
\title{Calculates posterior mean of a function of parameters}
\usage{
get_posterior_mean(X, ...)
}
\arguments{
\item{X}{either an array of posterior samples (either a parameter from \code{Posterior}, or
an object generated by \link{get_posterior_fun}), or the BSFG_state object with re-loaded Posterior}

\item{FUN}{(optional) if \code{X} is a BSFG_state object, the function to calculate the synthetic parameter}

\item{samples}{(optional) vector of posterior sample indexes to use for calculations}
}
\value{
posterior mean matrix
}
\description{
The synthetic parameter can be pre-calculated using \link{get_posterior_FUN}, or provided directly
    as an array with the first dimension the samples.
}
