% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{Image}
\alias{Image}
\title{Modification to \code{image()}.}
\usage{
Image(X, dimnames = FALSE, title = NULL, include_zero = TRUE, ...)
}
\arguments{
\item{colors}{}
}
\value{
ggplot object
}
\description{
Like `image,CHMfactor-method`, but with better defaults for correlation matrices especially.
}
\details{
The default is to have the colors evenly spaced, instead of having the full range <0 and >0.
}
