% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{initialize_variables_BSFG}
\alias{initialize_variables_BSFG}
\title{Initialize BSFG variables}
\usage{
initialize_variables_BSFG(BSFG_state, ...)
}
\arguments{
\item{BSFG_state}{BSFG_state object as returned by \link{setup_model_BSFG}}

\item{...}{Currenlty not supported}
}
\value{
BSFG_state object with current_state initialized with all variables
}
\description{
Initializes all variables in BSFG model with random draws.
Most variables are drawn from N(0,1) distributions, or uniformly for h2 parameters
}
