% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Posterior.R
\name{load_posterior_param}
\alias{load_posterior_param}
\title{load a set of posterior samples into a single array}
\usage{
load_posterior_param(folder = "Posterior", param)
}
\arguments{
\item{folder}{folder to find Posterior chunk files}

\item{param}{Name of parameter to load}
}
\value{
array with all chuncks of Posterior samples appended together
}
\description{
load a set of posterior samples into a single array
}
