% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{missing_data_model}
\alias{missing_data_model}
\title{Sample missing data}
\usage{
missing_data_model(Y, data_model_parameters, current_state = NULL)
}
\arguments{
\item{Y}{data matrix n_Y x p_Y}

\item{data_model_parameters}{List of parameters necessary for the data model.
Here, a Matrix of coordinates of NAs in Y}

\item{current_state}{current_state from BSFG_state}
}
\value{
Eta matrix of latent data: n x p
}
\description{
Function to sample missing data given model parameters. This is also a template
    for data_model functions.
}
\details{
Note: The model must return an appropriate Eta matrix, even if current_state is NULL
}
