% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{print.BSFG_state}
\alias{print.BSFG_state}
\title{Print statistics on current BSFG state}
\usage{
print.BSFG_state(BSFG_state)
}
\description{
Print statistics on current BSFG state
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{BSFG_init}}, \code{\link{summary.BSFG_state}}
}
