% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{reload_Posterior}
\alias{reload_Posterior}
\title{Re-loads a full Posterior list with all parameters}
\usage{
reload_Posterior(BSFG_state, params = NULL)
}
\arguments{
\item{BSFG_state}{a BSFG_state object (with empty Posterior)}

\item{params}{list of parameters to load. If NULL, all parameters will be loaded}
}
\value{
Posterior list, as part of a BSFG_state object
}
\description{
Re-loads a full Posterior list with all parameters
}
