% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{reorder_factors}
\alias{reorder_factors}
\title{Re-orders factors in decreasing order of magnitude}
\usage{
reorder_factors(BSFG_state, factor_order = NULL,
  drop_cor_threshold = 0.6)
}
\description{
Re-orders factors in decreasing order of magnitude
}
\seealso{
\code{\link{sample_BSFG}}, \code{\link{plot.BSFG_state}}
}
