% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_BSFG_functions.R
\name{sample_MME_fixedEffects}
\alias{sample_MME_fixedEffects}
\title{Sample from Mixed Model equations given a diagonal K matrix}
\usage{
sample_MME_fixedEffects(Y, W, Sigma_Choleskys, h2s_index, tot_Eta_prec,
  prior_mean, prior_prec, ncores)
}
\arguments{
\item{W}{matrix of size n x b}

\item{tot_Eta_prec}{the inverse of the total variance}

\item{prior_mean}{the vector \eqn{\mu}}

\item{prior_prec}{the diagonal of the matrix K^{-1}. All elements must be > 0}

\item{y}{vector of length n}

\item{Wp}{\eqn{R_perm * W} (see below)}

\item{Cholesky_R}{Cholesky decomposition of the sparse matrix R calculated by
\code{Cholesky(R_hat,perm=T,super=T)}. The factorization is  \eqn{P' L L' P}. R_hat here is
\eqn{R*tot_Eta_prec}, i.e., the covariance up to normalization by the total variance.}

\item{chol_R}{the matrix \eqn{P' L} corresponding to the Cholesky_R decomposition above}

\item{R_Perm}{either \code{NULL} if \eqn{P} is diagonal, or the \eqn{P} matrix.}
}
\description{
Sample from Mixed Model equations given a diagonal K matrix with sparse but non-diagonal R
}
\details{
This function draws a sample of the vector \eqn{theta} given the model:
    \deqn{y = W\theta + e}
    \deqn{\theta \sim N_b(\mu,K)} \deqn{e \sim N(0,R)} The algorithm follows that shown in the
    MCMCglmm course notes. This involves solving:
    \deqn{\tilde{\theta} = C^{-1}W'R^{-1}(1 - W\theta_{*} - e_{*})}
    \deqn{C = W'R^{-1}W + K^{-1}}
    \deqn{\theta_{*} \sim N(\mu,K), e_{*} \sim N(0,R)}

To speed up the calculations, \eqn{R_hat = R*tot_Eta_prec = P' L L' P} is pre-calculated, and used
in place of \eqn{R^{-1}}.
}
