% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Posterior.R
\name{save_posterior}
\alias{save_posterior}
\title{Saves a chunk of posterior samples}
\usage{
save_posterior(Posterior, folder = "Posterior", ID = 1)
}
\arguments{
\item{Posterior}{a Posterior list from a BSFG_state object}

\item{folder}{the folder to save the RData files}

\item{ID}{the ID of this chunck of samples. Should be consecutive with previous chunks}
}
\description{
Saves a chunk of posterior samples, each paramter in it's own RData file.
The complete chain for a particular parameter can be re-loaded with \code{load_posterior}
}
