% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{save_posterior_chunk}
\alias{save_posterior_chunk}
\title{Saves a chunk of posterior samples}
\usage{
save_posterior_chunk(BSFG_state, ID = 0)
}
\arguments{
\item{ID}{the ID of this chunck of samples. Should be consecutive with previous chunks}

\item{Posterior}{a Posterior list from a BSFG_state object}

\item{folder}{the folder to save the RData files}
}
\value{
Posterior
}
\description{
Saves a chunk of posterior samples, each paramter in it's own RData file.
The complete chain for a particular parameter can be re-loaded with \code{load_posterior}
}
