% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{save_posterior_samples}
\alias{save_posterior_samples}
\title{Saves current state in Posterior}
\usage{
save_posterior_samples(sp_num, current_state, Posterior)
}
\description{
Saves current state in Posterior
}
\seealso{
\code{\link{sample_BSFG}}, \code{\link{plot.BSFG_state}}
}
