% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{set_priors_BSFG}
\alias{set_priors_BSFG}
\title{Set priors for BSFG model.}
\usage{
set_priors_BSFG(BSFG_state, priors = BSFG_priors())
}
\arguments{
\item{BSFG_state}{BSFG_state object as returned by \link{setup_model_BSFG}}

\item{priors}{List as returned by \link{BSFG_priors}}
}
\value{
BSFG_state object with prior information added.
}
\description{
See \link{BSFG_priors} for more information
}
