% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{setup_model_BSFG}
\alias{setup_model_BSFG}
\title{Set up a BSFG model}
\usage{
setup_model_BSFG(Y, formula, extra_regressions = NULL, data, relmat = NULL,
  cis_genotypes = NULL, run_parameters = BSFG_control(),
  posteriorSample_params = c("Lambda", "U_F", "F", "delta", "tot_F_prec",
  "F_h2", "tot_Eta_prec", "resid_h2", "B1", "B2_F", "B2_R", "U_R",
  "cis_effects", "Lambda_m_eff"), posteriorMean_params = c(), setup = NULL,
  run_ID = "BSFG_run")
}
\arguments{
\item{Y}{either a) a n x p matrix of data (n individuals x p traits), or b) a list describing
the observation_model, data, and associated parameters. This list should contain:
i) \code{observation_model}: a function modeled after \code{missing_observation_model}
    (see code by typing this into the console) that draws posterior samples of Eta conditional on
    the observations (Y) and the current state of the BSFG model (current_state).
    The function should have the same form as \link{missing_data},
    and must return Eta even if current_state is NULL.
ii) \code{observations}: a data.frame containing the observaition-level data and associated covariates.
    This must include a column \code{ID} that is also present in \code{data}
iii) any other parameters necessary for \code{observation_model}}

\item{formula}{RHS of a model. The syntax is similar to \link{lmer}.
Random effects are specified by (1+factor | group), with the left side of the '|' a design
matrix, and the right side a random effect factor (group). For each random effect factor, a
covariance matrix (\code{K_mats}) or precision matrix (\code{K_inv_mats}) can be provided.
Unlike in \code{lmer}, each variable or covariate in the design matrix is given an
independent random effect (ie no covariance among random effects is modeled), so two bars '||'
gives an identical model to one bar.
Note: the speed of the model will decrease dramatically with the number of random effects (multiplied
by h2_divisions).
Fixed effects only apply to the model residuals (ie X1) below, and are not regularized (prior precision = 0)}

\item{extra_regressions}{Optional. A list including either:
i) the matrix X (n x b) of regression coeffients, or
ii) two matrices U (n x m) and V (m x b) such that X = U*V
also, logical variables \code{resids} and \code{fixed} specify whether these coefficients apply to either or
both of the model residuals or the factors}

\item{data}{data.frame with n rows containing columns corresponding to the fixed and random
effects}

\item{relmat}{Optional. A list of covariance matrices for random effects. If none provided
for any of the random effects, K is assumed to be the identity.}

\item{cis_genotypes}{Optional. A list of n x ci matrices of length p giving cis-effect coefficients for each trait}

\item{run_parameters}{See \link{BSFG_control}}

\item{posteriorSample_params}{A character vector giving names of parameters to save all posterior samples}

\item{posteriorMean_params}{A character vector giving names of parameters to save only the posterior mean.}

\item{setup}{optional - a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}

\item{run_ID}{A unique identifier for this model. The code will create a folder with this name to hold all
posterior samples and diagnostic information during the run.}
}
\value{
An object of class BSFG_state with components:

current_state: a list of parameters in the current iteration of the sampler

Posterior: a list of arrays of posterior samples

RNG current state of R's Random number generator (for re-starting chaings)

traitnames: vector of trait names (from colnames of Y)

run_parameters, run_variables, data_matrices, priors, simulation: input data and
  parameters
}
\description{
Sets up the BSFG model, selects starting values, and pre-calculates matrices for the GIBBS
sampler.
}
\details{
The first step in fitting a BSFG model. This function setups up the model matrices based on the
fixed and random effect formulas provided. This function must be followed by calls to
\link{set_priors_BSFG}, \link{initialize_variables_BSFG} and \link{initialize_BSFG}, before
the Gibbs sampler can be run with \link{sample_BSFG}.

The model is specified as:

y_i = g(eta_i)

Eta = rbind(eta_1,...,eta_n) = X1*B1 + X2_R*B2_R + F*t(Lambda) + Z*U_R + E_R

F = X2_F * B2_F + Z*U_F + E_F

For sampling, we reparameterize as:

Qt*Eta = Qt*X*B + Qt*F*t(Lambda) + Qt*ZL*U_R + Qt*E_R

Qt*F = Qt*X_F * B_F + Qt*ZL*U_F + Qt*E_F

where LTL = K and ZL = Z*L

We sample the quantities Qt*Eta, Qt*F, B, Lambda, U_R, U_F. We then back-calculate Eta and F.
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}}, \code{\link{plot.BSFG_state}}#'
}
