% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{summary.BSFG_state}
\alias{summary.BSFG_state}
\title{Print more detailed statistics on current BSFG state}
\usage{
summary.BSFG_state(BSFG_state)
}
\description{
Print more detailed statistics on current BSFG state
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{BSFG_init}}, \code{\link{print.BSFG_state}}
}
