% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{toDense}
\alias{toDense}
\title{Converts dense Matrix to "matrix" faster!}
\usage{
toDense(X)
}
\arguments{
\item{X}{a ddenseMatrix}
}
\value{
a matrix
}
\description{
Converts dense Matrix to "matrix" faster!
}
