% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{update_k}
\alias{update_k}
\title{Checks if factors (columns of Lambda) can be safely dropped}
\usage{
update_k(current_state, priors, run_parameters, data_matrices)
}
\description{
Factors are dropped if \code{prop} faction of the \lambda_{ij} elements are less than
\code{epsilon}. Checking happens stochastically at a decreasing rate during the chain controlled
by b0 and b1. If no factors can be dropped, then a new one is appended.
}
\seealso{
\code{\link{sample_BSFG}}, \code{\link{plot.BSFG_state}}
}
