% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_control}
\alias{BSFG_control}
\title{Set BSFG run parameters}
\usage{
BSFG_control(Posterior_folder = "Posterior", simulation = c(F, T),
  scale_Y = c(T, F), lambda_propto_Vp = TRUE, cauchy_sigma_tot = FALSE,
  b0 = 1, b1 = 5e-04, epsilon = 0.1, prop = 1, k_init = 20,
  h2_divisions = 100, h2_step_size = NULL, drop0_tol = 1e-14,
  K_eigen_tol = 1e-10, burn = 100, thin = 2,
  delta_iteractions_factor = 100, impute_missing = FALSE, svd_K = TRUE,
  ...)
}
\arguments{
\item{Posterior_folder}{path to folder to save posterior samples. Samples of each parameter
are saved in chuncks to limit memory requirements.}

\item{simulation}{Is this a fit to simulated data? If so, a setup list will be expected providing
the true values}

\item{scale_Y}{Should the Y values be centered and scaled? Recommend, except for simulated data.}

\item{lambda_propto_Vp}{Should the prior for lambda include tot_Eta_prec?}

\item{cauchy_sigma_tot}{Should the prior on sigma_j be half-cauchy?}

\item{b0}{parameter of the \code{update_k} function. See Bhattacharya and Dunson 2011}

\item{b1}{parameter of the \code{update_k} function. See Bhattacharya and Dunson 2011}

\item{epsilon}{parameter of the \code{update_k} function. Smallest \eqn{\lambda_{ij}} that is
considered "large", signifying a factor should be kept. See Bhattacharya and Dunson 2011}

\item{prop}{proportion of \eqn{\lambda{ij}} elements in a column of \eqn{\Lambda} that must be smaller than
\code{epsilon} before factor is dropped. See Bhattacharya and Dunson 2011}

\item{k_init}{initial number of factors}

\item{h2_divisions}{A scalar or vector of length equal to number of random effects. In BSFG, random
effects are re-scaled as percentages of the total variation. Then a discrete prior spanning [0,1)
with \code{h2_divisions} equally spaced values is constructred for each variance component. If
\code{h2_divisions} is a scalar, the prior for each variance component has this number of divisions.
In the joint prior over all variance components, combinations of variance components with total variance != 1
are assigned a prior of zero and ignored.}

\item{h2_step_size}{Either NULL, or a scaler in the range (0,1] giving specifying the range of h2 values for a Metropolis-Hastings
update step for each h2 parameter vector. If NULL, h2's will be sampled based on the marginal probability
over all possible h2 vectors. If a scalar, a Metropolis-Hastings update step will be used for each h2 vector.
The trail value will be selected uniformly from all possible h2 vectors within this Euclidean distance from the current vector.}

\item{drop0_tol}{A scalar giving the a tolerance for the \code{drop0()} function that will be applied
to various symmetric (possibly) sparse matrices to try to fix numerical errors and increase sparsity.}

\item{K_eigen_tol}{A scalar giving the minimum eigenvalue of a K matrix allowed. During pre-processing,
eigenvalues of each K matrix will be calculated using \code{svd(K)}. Only eigenvectors of K with corresponding eigenvalues
greater than this value will be kept. If smaller eigenvalues exist, the model will be transformed
to reduce the rank of K, by multiplying Z by the remaining eigenvectors of K. This transformation
is undone before posterior samples are recorded, so posterior samples of \code{U_F} and \code{U_R} are
untransformed.}

\item{burn}{burnin length of the MCMC chain}

\item{thin}{thinning rate of the MCMC chain}

\item{delta_iteractions_factor}{Number of times to iterate through sample_delta per iteration of the other parameters}

\item{impute_missing}{If FALSE, missing data will be ignored in the sampler. If TRUE, missing data will be imputed during sampling.}

\item{svd_K}{If TRUE, and a 1-random effect model is specified, the the diagonalization of ZKZt is accomplished using this algorithm:
https://math.stackexchange.com/questions/67231/singular-value-decomposition-of-product-of-matrices which doesn't require forming ZKTt.
If FALSE, the SVD of ZKZt is calculated directly. TRUE is generally faster if the same genomes are repeated several times.}
}
\description{
Function to create run_parameters list for initializing BSFG model
}
\seealso{
\code{\link{BSFG_init}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}}
}
