% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_init}
\alias{BSFG_init}
\title{Initialize a BSFG model}
\usage{
BSFG_init(Y, fixed, random, data, priors, run_parameters, A_mats = NULL,
  A_inv_mats = NULL, sampler = c("fast_BSFG", "general_BSFG"), ncores = 1,
  simulation = c(F, T), setup = NULL, verbose = T)
}
\arguments{
\item{Y}{a n x p matrix of data (n individuals x p traits)}

\item{fixed}{formula for the fixed effects}

\item{random}{formula for the random effects. Multiple random effects can be specified as
~random1+random2+..., but currently there is no support for interactions. Each random effect
must correspond to a column of data (which should be a factor)}

\item{data}{data.frame with n rows containing columns corresponding to the fixed and random
effects}

\item{priors}{list providing hyperparameters for the model priors. T}

\item{run_parameters}{list providing various parameters for the model run. See \link{BSFG_control}}

\item{A_mats}{list of covariance matrices for random effects. If none provided (and none provided
for A_inv_mats), assumed to be the identity.}

\item{A_inv_mats}{list of covariance matrices for random effects. If none provided (and none
provided for A_mats), assumed to be the identity.}

\item{ncores}{for \code{general_BSFG}, number of cores to use during initialization.}

\item{setup}{optional - a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}

\item{verbose}{should progress in initialization be reported?}
}
\value{
An object of class BSFG_state with components:

current_state: a list of parameters in the current iteration of the sampler

Posterior: a list of arrays of posterior samples

RNG current state of R's Random number generator (for re-starting chaings)

traitnames: vector of trait names (from colnames of Y)

run_parameters, run_variables, data_matrices, priors, simulation: input data and
  parameters
}
\description{
Sets up the BSFG model, selects starting values, and pre-calculates matrices for the GIBBS
sampler.
}
\details{
The first step in fitting a BSFG model. This function setups up the model matrices based on the
fixed and random effect formulas provided, initializes all of model parameters, and then
pre-calculates a set of matrices and transformations to speed up each iteration of the Gibbs
sampler.
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}}, \code{\link{plot.BSFG_state}}
}
