% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_init}
\alias{BSFG_init}
\title{Initialize a BSFG model}
\usage{
BSFG_init(Y, model, data, priors, run_parameters, A_mats = NULL,
  A_inv_mats = NULL, data_model = "missing_data",
  data_model_parameters = NULL, sampler = c("fast_BSFG", "general_BSFG"),
  ncores = detectCores(), simulation = c(F, T), setup = NULL,
  verbose = T)
}
\arguments{
\item{Y}{a n x p matrix of data (n individuals x p traits)}

\item{model}{RHS of a model. The syntax is similar to \link{lmer}.
Random effects are specified by (1+factor | group), with the left side of the '|' a design
matrix, and the right side a random effect factor (group). For each random effect factor, a
covariance matrix (\code{A_mats}) or precision matrix (\code{A_inv_mats}) can be provided.
Unlike in \code{lmer}, each variable or covariate in the design matrix is given an
independent random effect (ie no covariance among random effects is modeled), so two bars '||'
gives an identical model to one bar.
Note: the speed of the model will decrease dramatically with the number of random effects (multiplied
by h2_divisions)}

\item{data}{data.frame with n rows containing columns corresponding to the fixed and random
effects}

\item{priors}{list providing hyperparameters for the model priors. T}

\item{run_parameters}{list providing various parameters for the model run. See \link{BSFG_control}}

\item{A_mats}{list of covariance matrices for random effects. If none provided (and none provided
for A_inv_mats) for any of the random effects, A is assumed to be the identity.}

\item{A_inv_mats}{list of precision matrices for random effects. If none provided (and none
provided for A_mats) for any of the random effects, A is assumed to be the identity.}

\item{data_model}{either a character vector identifying a provided data_model
(ex. 'missing_data' calls \link{missing_data_model}),
or a function modeled after \code{missing_data_model} (see code by typing this into the console)
that draws posterior samples of Eta conditional on
the observations (Y) and the current state of the BSFG model (current_state). The function should have
the same form as \link{missing_data}, and must return Eta even if current_state is NULL}

\item{data_model_parameters}{a list of parameters necessary for computing \code{data_model}.
Ex. Y_missing for the missing_data model.
If \code{data_model == 'missing_data'}, Y_missing is computed automatically.}

\item{ncores}{for \code{general_BSFG}, number of cores to use during initialization.}

\item{setup}{optional - a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}

\item{verbose}{should progress in initialization be reported?}
}
\value{
An object of class BSFG_state with components:

current_state: a list of parameters in the current iteration of the sampler

Posterior: a list of arrays of posterior samples

RNG current state of R's Random number generator (for re-starting chaings)

traitnames: vector of trait names (from colnames of Y)

run_parameters, run_variables, data_matrices, priors, simulation: input data and
  parameters
}
\description{
Sets up the BSFG model, selects starting values, and pre-calculates matrices for the GIBBS
sampler.
}
\details{
The first step in fitting a BSFG model. This function setups up the model matrices based on the
fixed and random effect formulas provided, initializes all of model parameters, and then
pre-calculates a set of matrices and transformations to speed up each iteration of the Gibbs
sampler.
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}}, \code{\link{plot.BSFG_state}}
}
