% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_init}
\alias{BSFG_init}
\title{Initialize a BSFG model}
\usage{
BSFG_init(Y, model, data, factor_model_fixed = NULL, priors = BSFG_priors(),
  run_parameters = BSFG_control(), K_mats = NULL, K_inv_mats = NULL,
  QTL_resid = NULL, QTL_factors = NULL, cis_genotypes = NULL,
  Sigma_Choleskys = NULL, randomEffect_C_Choleskys = NULL,
  invert_aI_bZKZ = NULL, invert_aZZt_Kinv = NULL,
  posteriorSample_params = c("Lambda", "U_F", "F", "delta", "tot_F_prec",
  "F_h2", "tot_Eta_prec", "resid_h2", "B", "B_F", "U_R", "cis_effects"),
  posteriorMean_params = c(), ncores = detectCores(), setup = NULL,
  verbose = T)
}
\arguments{
\item{Y}{either a) a n x p matrix of data (n individuals x p traits), or b) a list describing
the observation_model, data, and associated parameters. This list should contain:
i) \code{observation_model}: a function modeled after \code{missing_observation_model}
    (see code by typing this into the console) that draws posterior samples of Eta conditional on
    the observations (Y) and the current state of the BSFG model (current_state).
    The function should have the same form as \link{missing_data},
    and must return Eta even if current_state is NULL.
ii) \code{observations}: a data.frame containing the observaition-level data and associated covariates.
    This must include a column \code{ID} that is also present in \code{data}
iii) any other parameters necessary for \code{observation_model}}

\item{model}{RHS of a model. The syntax is similar to \link{lmer}.
Random effects are specified by (1+factor | group), with the left side of the '|' a design
matrix, and the right side a random effect factor (group). For each random effect factor, a
covariance matrix (\code{K_mats}) or precision matrix (\code{K_inv_mats}) can be provided.
Unlike in \code{lmer}, each variable or covariate in the design matrix is given an
independent random effect (ie no covariance among random effects is modeled), so two bars '||'
gives an identical model to one bar.
Note: the speed of the model will decrease dramatically with the number of random effects (multiplied
by h2_divisions)}

\item{data}{data.frame with n rows containing columns corresponding to the fixed and random
effects}

\item{factor_model_fixed}{Fixed effect model formula specific to the latent factors. Optional. If NULL,
the fixed effects for the latent factors will be the same as for Eta, except that the columns will be centered.
If a formula is provided (no random effects allowed), the model will be applied to
each factor, again with the intercept dropped (or set to zero). Note: Random effects in \code{model}
are applied to both Eta and F.}

\item{priors}{list providing hyperparameters for the model priors. This must include: for \code{fixed_var},
\code{tot_Eta_var}, and \code{tot_F_var}, a list of \code{V} and \code{nu}, which specify an inverse-gamma
distribution as in MCMCglmm. For \code{delta_1} and \code{delta_2} a list of \code{shape} and \code{rate} for
a gamma distribution. For \code{Lambda_df}, the degrees of freedom of the implied t-distribution.
Discrete priors on \code{h2_priors_factors} and \code{h2_priors_resids} are also required for \code{sample_BSFG},
but can be appended to \code{BSFG_state$priors} after running \code{BSFG_init}. This is often easier if
the total number of h2 divisions is not known beforehand (ie when multiple random effects are used).}

\item{run_parameters}{list providing various parameters for the model run. See \link{BSFG_control}}

\item{K_mats}{list of covariance matrices for random effects. If none provided (and none provided
for K_inv_mats) for any of the random effects, K is assumed to be the identity.}

\item{K_inv_mats}{list of precision matrices for random effects. If none provided (and none
provided for K_mats) for any of the random effects, K is assumed to be the identity.}

\item{QTL_resid}{a design matrix. Intended for marker genotypes. Corresponding coefficients
will be modeled with the \code{QTL_resid_var} priors. Applies to factor residuals \eqn{Y - F\Lambda^T}}

\item{QTL_factors}{a design matrix. Intended for marker genotypes. Corresponding coefficients
will be modeled with the \code{QTL_factor_var} priors. Applies to factors \eqn{F}.}

\item{cis_genotypes}{a list of design matrices of length \code{p} (ie number of columns of \code{Eta})
This is used to specify trait-specific fixed effects, such a cis-genotypes}

\item{Sigma_Choleskys}{Pre-calculated matrices from \code{BSFG_state$run_variables$Sigma_Choleskys}
can be provide directly. For general_BSFG sampler.}

\item{randomEffect_C_Choleskys}{Pre-calculated matrices from \code{BSFG_state$run_variables$randomEffect_C_Choleskys}
can be provide directly. For general_BSFG sampler.}

\item{invert_aI_bZKZ}{Pre-calculated matrices from \code{BSFG_state$run_variables$invert_aI_bZKZ}
can be provide directly. For fast_BSFG sampler.}

\item{invert_aZZt_Kinv}{Pre-calculated matrices from \code{BSFG_state$run_variables$invert_aZZt_Kinv}
can be provide directly. For fast_BSFG sampler.#'}

\item{posteriorSample_params}{A character vector giving names of parameters to save all posterior samples}

\item{posteriorMean_params}{A character vector giving names of parameters to save only the posterior mean.}

\item{ncores}{for \code{general_BSFG}, number of cores to use during initialization.}

\item{setup}{optional - a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}

\item{verbose}{should progress in initialization be reported?}
}
\value{
An object of class BSFG_state with components:

current_state: a list of parameters in the current iteration of the sampler

Posterior: a list of arrays of posterior samples

RNG current state of R's Random number generator (for re-starting chaings)

traitnames: vector of trait names (from colnames of Y)

run_parameters, run_variables, data_matrices, priors, simulation: input data and
  parameters
}
\description{
Sets up the BSFG model, selects starting values, and pre-calculates matrices for the GIBBS
sampler.
}
\details{
The first step in fitting a BSFG model. This function setups up the model matrices based on the
fixed and random effect formulas provided, initializes all of model parameters, and then
pre-calculates a set of matrices and transformations to speed up each iteration of the Gibbs
sampler.

The model is specified as:

y_i = g(eta_i)
Eta = rbind(eta_1,...,eta_n) = X*B + F*t(Lambda) + Z*U_R + E_R
F = X_F * B_F + Z*U_F + E_F

For sampling, we reparameterize as:

Qt*Eta = Qt*X*B + Qt*F*t(Lambda) + Qt*Z*U_R + Qt*E_R
Qt*F = Qt*X_F * B_F + Qt*Z*U_F + Qt*E_F

We sample the quantities Qt*Eta, Qt*F, B, Lambda, U_R, U_F. We then back-calculate Eta and F.
}
\seealso{
\code{\link{BSFG_control}}, \code{\link{sample_BSFG}}, \code{\link{print.BSFG_state}}, \code{\link{plot.BSFG_state}}#'
}
