% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_priors}
\alias{BSFG_priors}
\title{Set BSFG priors}
\usage{
BSFG_priors(tot_Y_var = list(V = 0.5, nu = 3), tot_F_var = list(V = 18/20,
  nu = 20), h2_priors_resids_fun = function(h2s, n) 1,
  h2_priors_factors_fun = function(h2s, n) 1, Lambda_prior = list(sampler =
  sample_Lambda_prec_ARD, Lambda_df = 3, delta_1 = list(shape = 2, rate = 1),
  delta_2 = list(shape = 3, rate = 1), delta_iteractions_factor = 100),
  B_prior = list(sampler = sample_B_prec_ARD, global = list(V = 1, nu = 3),
  global_F = list(V = 1, nu = 3), B_df = 3, B_F_df = 3),
  QTL_prior = list(sampler = sample_QTL_prec_horseshoe, separate_QTL_shrinkage
  = T, cauchy_iteractions_factor = 10))
}
\arguments{
\item{tot_Y_var}{List of parameters of inverse gamma distribution for residual variances, specifically:
\code{V} and \code{nu}, give shape = \code{nu/2} and scale = \code{nu*V/2}, so mean = \code{\frac{V*nu}{nu-2}}}

\item{tot_F_var}{List of parameters of inverse gamma distribution for factor variances. See \code{tot_Y_var}.
This parameter provides the parameter extension of Ghosh and Dunson (2009), but is removed
from all Factor parameters before they are saved in Posterior}

\item{h2_priors_resids_fun}{function for that returns prior probability for a given value of h2
for each random effect. Should take two argument - a vector \code{h2} values for each random effect,
and \code{n} - the number of discrete levels of the prior.
Alternatively, can be a scalar or vector of (relative) prior values for each value of the
discrete prior.}

\item{h2_priors_factors_fun}{see \code{h2_priors_resids_fun}. Same, but for the h2s of the factors.}

\item{Lambda_prior}{A list with elements:
1) \code{sampler}: a function that draws samples of the precision matrix for Lambda. Ex: \code{sample_Lambda_prec_ARD}; 2)
any other hyperparameters and control parameters for \code{sampler}}

\item{B_prior}{A list with elements:
1) \code{sampler}: a function that draws samples of the precision matrix for B and B_F Ex: \code{sample_B_prec_ARD}; 2)
any other hyperparameters and control parameters for \code{sampler}}

\item{QTL_prior}{A list with elements:
1) \code{sampler}: a function that draws samples of the precision matrix for B_QTL and B_QTL_F Ex: \code{sample_QTL_prec_horseshoe}; 2)
any other hyperparameters and control parameters for \code{sampler}}
}
\value{
a list with each of the prior components specified above.
}
\description{
Function to create list of priors for BSFG model.
}
\details{
Default values are provided, but any can be replaced. Note: \code{h2_priors_resids} and
    \code{h2_priors_factors} can be set after calling \code{BSFG_init} and before \code{sample_BSFG}
    if that is easier.
}
