% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{BSFG_priors}
\alias{BSFG_priors}
\title{Set BSFG priors}
\usage{
BSFG_priors(fixed_var = list(V = 1, nu = 3), fixed_resid_var = NULL,
  fixed_factors_var = NULL, tot_Y_var = list(V = 0.5, nu = 3),
  tot_F_var = list(V = 18/20, nu = 20), delta_1 = list(shape = 2.1, rate =
  1/20), delta_2 = list(shape = 3, rate = 1), Lambda_df = 3, B_df = 3,
  B_F_df = 3, h2_priors_resids_fun = function(h2s, n) 1,
  h2_priors_factors_fun = function(h2s, n) 1)
}
\arguments{
\item{fixed_var}{List of parameters of inverse gamma distribution for fixed effects, specifically:
\code{V} and \code{nu}, give shape = \code{nu/2} and scale = \code{nu*V/2}, so mean = \code{\frac{V*nu}{nu-2}}
Will be applied to fixed effects of residuals (\code{B}) and factors (\code{B_F}).}

\item{fixed_resid_var}{If provided, overides \code{fixed_var} for fixed effects of residuals (\code{B}).}

\item{fixed_factors_var}{If provided, overides \code{fixed_var} for fixed effects of factors (\code{B_F}).}

\item{tot_Y_var}{List of parameters of inverse gamma distribution for residual variances. See \code{fixed_var}.}

\item{tot_F_var}{List of parameters of inverse gamma distribution for factor variances. See \code{fixed_var}.
This parameter provides the parameter extension of Ghosh and Dunson (2009), but is removed
from all Factor parameters before they are saved in Posterior}

\item{delta_1}{List of parameters of inverse gamma distribution for \code{delta_1}. Specifically:
\code{shape} and \code{rate}. This parameter is the column-shrinkage of the first factor.}

\item{delta_2}{List of parameters of inverse gamma distribution for \code{delta_2 \dots delta_k}.
Specifically: \code{shape} and \code{rate}.
This is provides the additional column-shrinkage of higher-order columns.}

\item{Lambda_df}{Degrees of freedom of individual parameter shrinkage of Lambda from implied
t-distribution}

\item{B_df}{Degrees of freedom of individual parameter shrinkage of B from implied
t-distribution}

\item{B_F_df}{Degrees of freedom of individual parameter shrinkage of B_F from implied
t-distribution}

\item{h2_priors_resids_fun}{function for that returns prior probability for a given value of h2
for each random effect. Should take two argument - a vector \code{h2} values for each random effect,
and \code{n} - the number of discrete levels of the prior.
Alternatively, can be a scalar or vector of (relative) prior values for each value of the
discrete prior.}

\item{h2_priors_factors_fun}{see \code{h2_priors_resids_fun}. Same, but for the h2s of the factors.}
}
\value{
a list with each of the prior components specified above.
}
\description{
Function to create list of priors for BSFG model.
}
\details{
Default values are provided, but any can be replaced. Note: \code{h2_priors_resids} and
    \code{h2_priors_factors} can be set after calling \code{BSFG_init} and before \code{sample_BSFG}
    if that is easier.
}
