% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_functions.R
\name{load_posterior_param}
\alias{load_posterior_param}
\title{load a set of posterior samples into a single array}
\usage{
load_posterior_param(BSFG_state, param)
}
\arguments{
\item{param}{Name of parameter to load}

\item{folder}{folder to find Posterior chunk files}
}
\value{
array with all chuncks of Posterior samples appended together
}
\description{
load a set of posterior samples into a single array
}
