% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSFG_master.R
\name{missing_data_model}
\alias{missing_data_model}
\title{Sample missing data}
\usage{
missing_data_model(Y, data_model_parameters, current_state = list(),
  data_matrices = NULL)
}
\arguments{
\item{Y}{data matrix n_Y x p_Y}

\item{data_model_parameters}{List of parameters necessary for the data model.
Here, a Matrix of coordinates of NAs in Y}

\item{current_state}{current_state from BSFG_state}

\item{data_matrices}{from BSFG_state}
}
\value{
list of data_model variables including:

Eta matrix of latent data: n x p
}
\description{
Function to sample missing data given model parameters.
}
\details{
This is also a template for \code{data_model} functions.

The function should draw a posterior sample for Eta (the (potentially) latent
   data on the linear scale) given the factor model and the observed data. It returns a list
   with Eta and any other variables associated with the data_model.
   These variables are added to current_state, but are not currently saved in Posterior

Initial values, hyperparameters, and any helper functions / parameters can be passed
   in \code{data_model_parameters}.

When run with an empty \code{current_state} and \code{data_matrices == NULL}, it should return
   a matrix Eta of the correct dimensions, but the values are unimportant.
}
