% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Posterior.R
\name{reload_Posterior}
\alias{reload_Posterior}
\title{Re-loads a full Posterior list with all parameters}
\usage{
reload_Posterior(Posterior, folder)
}
\arguments{
\item{Posterior}{an empty Posterior list (after call to \link{clear_Posterior})}

\item{folder}{folder with all posterior chunk files}
}
\value{
Posterior list, as part of a BSFG_state object
}
\description{
Re-loads a full Posterior list with all parameters
}
